updated twilio and eta middleware
